[JIT] Enable @ignore and @unused syntax for ignoring properties
**Summary**
This commit enables `@ignore` and `@unused` syntax for ignoring
properties, but works only if these decorators are applied to a function
before the @property decorator. Despite this, ignoring properties is
more intuitive with this feature enabled.

**Test Plan**
This commit updates the existing unit tests for class type and module
properties to test properties ignored using `@ignore` and `@unused`.
Summary:

ghstack-source-id: 77d06f304551bd3f63b5b451fa2856c2ba0bfdae
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45261